refactor: move SpringDoc PropertyCustomiser to commons
