+++ commons/spring/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
 org.elaastix.commons.jpa.ElaastixHibernateAutoConfiguration
+org.elaastix.commons.openapi.SpringdocAutoConfiguration
